Update host configurations and add new files for ignition, netplan, and Nginx setup
In this commit, the ansible host configurations are updated. New configuration files were added for ignition which includes setup for storage and user details, for netplan which includes network setup, and for Nginx which involves setup for Docker and default server details. The .gitignore file was also updated to ignore .idea files.